Update include string match
Regex matching for a literal `*` on a word boundary is not working the
way I would like.  This is much easier in bash with the following:
```sh
[[ "${includestr}" == *"*"* ]]
```

Change the wildward match to bash instead of perl regex.  It is just for
confirmation anyway; nothing is gained by using a regex here.